move db.queries module to deprecated namespace
