+++ db/tests/deprecated/queries/test_base.py
@@ -1,5 +1,5 @@
 from sqlalchemy import inspect
-from db.queries.base import DBQuery, InitialColumn, JoinParameter
+from db.deprecated.queries.base import DBQuery, InitialColumn, JoinParameter
 from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.transforms import base as tbase
 from db.deprecated.metadata import get_empty_metadata
