+++ db/tests/transforms/test_finish_specifying_summarization.py
@@ -3,7 +3,7 @@ from sqlalchemy import inspect
 import types
 import functools
 
-from db.queries.base import DBQuery, InitialColumn, JoinParameter
+from db.deprecated.queries.base import DBQuery, InitialColumn, JoinParameter
 from db.deprecated.metadata import get_empty_metadata
 from db.transforms.base import Summarize, SelectSubsetOfColumns, Limit
 from db.deprecated.columns import get_column_attnum_from_name
